Added PipelineSetId() and PipelineGetId(). This can be used by a client to prevent race conditions when receiving signals from gstd.
E.g.
1. A client starts playback of an audio file
2. Gstreamer reaches the end of the audio file and gstd sends an eos-signal to the client
3. Before the client receives the eos-signal, the client decides to play another file using the same pipe. It sets the pipe back to READY, changes the filename and sets the pipe to PLAYING again
4. Now the eos-signal from the first audio file is received by the client application
5. The client will handle the eos-signal, e.g. it sets the pipe back to READY and seeks to position 0 ns
-> This will accidently stop the playback of the second audio file

To prevent those race conditions, each pipe has an id, which can be changed by the client.
This id is also included in each signal as first parameter.
To prevent the race condition shown above, the client has to change the id before it sets the pipe to PLAYING again.
If the client receives the eos-signal from the first audio file it can decide to drop it, because the first argument in the signal has a different id.